implemented dao methods (#431)
* implemented dao methods

* changed specs and refactored

* consistent variable names

* refactored getAllTeamEventAttendance

* refactored deleteAllTEAttendance

* moved deleteCollection to firebase-utils

* deleted export default in firebase-utils

* fixed imports
