+++ backend/scripts/update-dev-db.ts
@@ -1,5 +1,5 @@
 import admin from 'firebase-admin';
-import configureAccount from '../src/utils/firebase-utils';
+import { configureAccount, deleteCollection } from '../src/utils/firebase-utils';
 
 require('dotenv').config();
 
@@ -13,47 +13,6 @@ type DbData = {
     data: unknown;
   }[];
 }[];
-
-async function deleteCollection(
-  db: FirebaseFirestore.Firestore,
-  collectionPath: string,
-  batchSize: number
-) {
-  const collectionRef = db.collection(collectionPath);
-  const query = collectionRef.orderBy('__name__').limit(batchSize);
-
-  return new Promise((resolve, reject) => {
-    deleteQueryBatch(db, query, resolve).catch(reject);
-  });
-}
-
-async function deleteQueryBatch(
-  db: FirebaseFirestore.Firestore,
-  query: FirebaseFirestore.Query,
-  resolve
-) {
-  const snapshot = await query.get();
-
-  const batchSize = snapshot.size;
-  if (batchSize === 0) {
-    // When there are no documents left, we are done
-    resolve();
-    return;
-  }
-
-  // Delete documents in a batch
-  const batch = db.batch();
-  snapshot.docs.forEach((doc) => {
-    batch.delete(doc.ref);
-  });
-  await batch.commit();
-
-  // Recurse on the next process tick, to avoid
-  // exploding the stack.
-  process.nextTick(() => {
-    deleteQueryBatch(db, query, resolve);
-  });
-}
 
 const readDbData = async (prodDb: FirebaseFirestore.Firestore): Promise<DbData> => {
   const allCollections = await prodDb.listCollections();
