+++ backend/src/utils/firebase-utils.ts
@@ -1,5 +1,6 @@
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/explicit-module-boundary-types
-const configureAccount = (sa: any, isProd: boolean) => {
+export const configureAccount = (sa: any, isProd: boolean) => {
   const configAcc = sa;
   let parsedPK;
   try {
@@ -16,4 +17,43 @@ const configureAccount = (sa: any, isProd: boolean) => {
   return configAcc;
 };
 
-export default configureAccount;
+export async function deleteCollection(
+  db: FirebaseFirestore.Firestore,
+  collectionPath: string,
+  batchSize: number
+) {
+  const collectionRef = db.collection(collectionPath);
+  const query = collectionRef.orderBy('__name__').limit(batchSize);
+
+  return new Promise((resolve, reject) => {
+    deleteQueryBatch(db, query, resolve).catch(reject);
+  });
+}
+
+export async function deleteQueryBatch(
+  db: FirebaseFirestore.Firestore,
+  query: FirebaseFirestore.Query,
+  resolve
+) {
+  const snapshot = await query.get();
+
+  const batchSize = snapshot.size;
+  if (batchSize === 0) {
+    // When there are no documents left, we are done
+    resolve();
+    return;
+  }
+
+  // Delete documents in a batch
+  const batch = db.batch();
+  snapshot.docs.forEach((doc) => {
+    batch.delete(doc.ref);
+  });
+  await batch.commit();
+
+  // Recurse on the next process tick, to avoid
+  // exploding the stack.
+  process.nextTick(() => {
+    deleteQueryBatch(db, query, resolve);
+  });
+}
